Move (binary) text relocation into own class
Also drop threaded text replacement, since it does not help
significantly (relocation got faster, bottleneck is decompression). It
would be better to parallellize over packages, instead of over files per
package.

Small improvement with separate classes for text replacement is that we
now compile the regex in the constructor; previously it was compiled per
binary to be relocated.
